backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pypdf==5.1.0
tiktoken==0.8.0
sentence-transformers==3.3.1
huggingface-hub==0.27.1
faiss-cpu==1.9.0.post1
numpy==2.2.1
python-dotenv==1.0.1
pydantic-settings==2.7.1
tenacity==9.0.0
python-multipart==0.0.20
torch>=2.1.0
pytest==8.3.4
